+++ forui/lib/src/widgets/scaffold.dart
@@ -56,6 +56,11 @@ class FScaffold extends StatelessWidget {
   /// ```
   final FToasterStyle? toasterStyle;
 
+  /// The axis in which to swipe to dismiss a toast. Defaults to [Axis.horizontal].
+  ///
+  /// Set to null to disable swiping to dismiss.
+  final Axis? toasterSwipeToDismiss;
+
   /// The main content area of the scaffold.
   final Widget child;
 
@@ -85,6 +90,7 @@ class FScaffold extends StatelessWidget {
     required this.child,
     this.scaffoldStyle,
     this.toasterStyle,
+    this.toasterSwipeToDismiss = Axis.horizontal,
     this.header,
     this.sidebar,
     this.footer,
@@ -108,6 +114,7 @@ class FScaffold extends StatelessWidget {
     return FSheets(
       child: FToaster(
         style: toasterStyle ?? context.theme.toasterStyle,
+        swipeToDismiss: toasterSwipeToDismiss,
         child: Row(
           children: [
             if (sidebar != null) ColoredBox(color: style.sidebarBackgroundColor, child: sidebar),
@@ -140,6 +147,7 @@ class FScaffold extends StatelessWidget {
     properties
       ..add(DiagnosticsProperty('style', scaffoldStyle))
       ..add(DiagnosticsProperty('toasterStyle', toasterStyle))
+      ..add(EnumProperty('toasterSwipeToDismiss', toasterSwipeToDismiss, defaultValue: Axis.horizontal))
       ..add(FlagProperty('childPad', value: childPad, ifTrue: 'contentPad', defaultValue: true))
       ..add(
         FlagProperty(